Replace the ttl::dim cpo with ttl::extent, and implement extent mapping.
The use of "extent" is more-or-less consistent with C++'s use of std::extent,
except that in ttl extent is a CPO class template along the lines of std::get.

We also have cases where we have one index space and we need to map that space
before looking up an extent, so we implemente map_extents for that.