Add the delta function.
The delta function complicates everything because, while it is a tensor
expression, it has neither a fixed scalar_type, nor fixed extents.

For the scalar type, it should produce either the identity or annihilator
element appropriate for its position in the tree. I implement this using `int`
for now.

For the extents, whenever we contract an index from the the delta function, the
paired index must contain bounded extents. I've decided to label the delta
function indices as "SYNTHETIC" and use the index type to propagate this
information up the tree.